% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{address}
\alias{address}
\title{Title Address Retriver}
\usage{
address(latlong = NULL)
}
\arguments{
\item{latlong}{A character string that includes the latitude and the longitude.}
}
\value{
A dataframe containing full address and coordinates.
}
\description{
The 'address_lookup' function is a function that provides the user the address of a given
a set of coordinates The funtion uses the google geocode API to retrieve the result.
}
\examples{
address(latlong = "58.3978, 15.5760")

}
