% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{check_for_retriever}
\alias{check_for_retriever}
\title{Check if retriever is on the user's path}
\usage{
check_for_retriever(...)
}
\description{
Check if retriever is on the user's path
}
\keyword{internal}

