% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLbyDisease.R
\docType{data}
\name{CLbyDisease}
\alias{CLbyDisease}
\title{CLbyDisease}
\format{
A list of 14 cancer types and the cell lines used as surrogates by the LINCS-L1000 project.
}
\usage{
data(CLbyDisease)
}
\description{
CLbyDisease
}
