% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Title Coordinates Retriver}
\usage{
coordinates(address = NULL)
}
\arguments{
\item{address}{A character string that includes address components}
}
\value{
A dataframe cointaining Full address and the coordinates
}
\description{
The coodinate function uses the google geocode API to retriever coordinates given address.
}
\examples{
coordinates(address="Master Mattias vag 7, Linkoping, Sweden")

}
