% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Creates a corpus  that can be used with create_locations}
\usage{
create_corpus(folder)
}
\arguments{
\item{folder}{Path to files}
}
\description{
Creates a corpus  that can be used with create_locations
}
\examples{
create_corpus()
}
\keyword{creation}
