% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus_html.R
\name{create_corpus_html}
\alias{create_corpus_html}
\title{HTML version of create_corpus, to be removed}
\usage{
create_corpus_html(data)
}
\arguments{
\item{data}{Results of prep_htmls}
}
\description{
HTML version of create_corpus, to be removed
}
\examples{
data <- create_corpus_html(data)
}
\keyword{wrapper}
