% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df.R
\name{create_df}
\alias{create_df}
\title{A function to create a paragraph-per-row dataframe inheriting folder-level metadata}
\usage{
create_df(folders, main)
}
\arguments{
\item{folders}{Path to .txt files from working directory}

\item{main}{Defaults to true. If true, attempts to find page numbers in the .txt names}
}
\description{
A function to create a paragraph-per-row dataframe inheriting folder-level metadata
}
\examples{
create_df()
}
\keyword{creation}
