% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_locations.R
\name{create_locations}
\alias{create_locations}
\title{Create neural embedding weights for each paragraph in a corpus}
\usage{
create_locations(corpus, embedding = "embeddings.bin")
}
\arguments{
\item{corpus}{Dataframe returned from make_corpus}

\item{embedding}{Path to neural embedding (.bin)}
}
\description{
Create neural embedding weights for each paragraph in a corpus
}
\examples{
create_locations()
}
\keyword{creation}
