% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Query a corpus for a topic and create a report summarizing and containing the results}
\usage{
create_report(country = NULL, query, data,
  embeddings = "embeddings.bin", locations = "embeddings.rds",
  type = "html", interactive = TRUE, thresh = 0.55, class = "pdf")
}
\arguments{
\item{country}{Country, must match the groupings from folders}

\item{query}{Query, in the form of "this is a query"}

\item{data}{Dataframe}

\item{embeddings}{Path to neural embedding (.bin)}

\item{locations}{Path to locations exported from create_locations}
}
\description{
Query a corpus for a topic and create a report summarizing and containing the results
}
\examples{
create_report()
}
\keyword{report}
