% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtualenv.R
\name{create_virtualenv}
\alias{create_virtualenv}
\title{Creates a Python 2.7 virtual environment for running OCR}
\usage{
create_virtualenv(name = "RetrieveR", type = "conda")
}
\arguments{
\item{name}{Name for the virtual environment, defaults to RetrieveR}
}
\description{
Creates a Python 2.7 virtual environment for running OCR
}
\examples{
create_virtualenv()
}
\keyword{creation}
