% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_embeddings.R
\name{download_embeddings}
\alias{download_embeddings}
\title{Download neural embeddings pre-trained on 1,000+ environmental policy documents from 40 nations and 50+ NGOs.}
\usage{
download_embeddings(path = "embeddings.bin")
}
\arguments{
\item{path}{Location to save the downloaded embeddings to, defaults to "/embeddings.bin"}
}
\description{
Download neural embeddings pre-trained on 1,000+ environmental policy documents from 40 nations and 50+ NGOs.
}
\examples{
download_embeddings()
}
\keyword{creation}
