% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_example.R
\name{download_example}
\alias{download_example}
\title{Downloads example PDFs for use with the vignette}
\usage{
download_example()
}
\description{
Downloads example PDFs for use with the vignette
}
\examples{
download_example()
}
\keyword{creation}
