% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{get_updates}
\alias{get_updates}
\title{Update the retriever's dataset scripts to the most recent versions.}
\usage{
get_updates()
}
\description{
This function will check if the version of the retriever's scripts in your local
directory \file{~/.retriever/scripts/} is up-to-date with the most recent official
retriever release. Note it is possible that even more updated scripts exist
at the retriever repository \url{https://github.com/weecology/retriever/tree/master/scripts}
that have not yet been incorperated into an official release, and you should
consider checking that page if you have any concerns.
}
\examples{
\donttest{
rdataretriever::get_updates()
}
}
\keyword{utilities}

