% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mac.R
\name{install_mac}
\alias{install_mac}
\title{Install the wordVectors package from source}
\usage{
install_mac()
}
\description{
Install the wordVectors package from source
}
\examples{
install_mac()
}
\keyword{creation}
