% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_windows.R
\name{install_windows}
\alias{install_windows}
\title{Install the wordVectors package from binary}
\usage{
install_windows()
}
\description{
Install the wordVectors package from binary
}
\examples{
install_windows()
}
\keyword{creation}
