% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corpus.R
\name{make_corpus}
\alias{make_corpus}
\title{Creates a corpus  that can be used with create_locations}
\usage{
make_corpus(folder)
}
\arguments{
\item{folder}{Path to files}
}
\description{
Creates a corpus  that can be used with create_locations
}
\examples{
make_corpus()
}
\keyword{creation}
