% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_htmls.R
\name{prep_htmls}
\alias{prep_htmls}
\title{Reads in htmls to a data frame}
\usage{
prep_htmls(path)
}
\arguments{
\item{path}{Path to the folder containing PDFs}
}
\description{
Reads in htmls to a data frame
}
\examples{
data <- prep_htmls("path/to/folder")

}
\keyword{wrapper}
