% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_wordvec.R
\name{prep_wordvec}
\alias{prep_wordvec}
\title{Corpus creating and data cleaning to create a .csv that can be fed into create_wordvec}
\usage{
prep_wordvec(folder)
}
\arguments{
\item{corpus}{Dataframe returned from make_corpus}

\item{embedding}{Path to neural embedding (.bin)}
}
\description{
Corpus creating and data cleaning to create a .csv that can be fed into create_wordvec
}
\examples{
prep_wordvec()
}
\keyword{creation}
