% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{print.update_log}
\alias{print.update_log}
\title{print the output from get_updates}
\usage{
\method{print}{update_log}(x, ...)
}
\description{
print the output from get_updates
}
\keyword{internal}

