% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever-package.R
\docType{package}
\name{rdataretriever}
\alias{rdataretriever}
\alias{rdataretriever-package}
\title{rdataretriever: R interface to the Data Retriever.}
\description{
rdataretriever: R interface to the Data Retriever.
}
\section{About}{


The Data Retriever automates the tasks of finding, downloading, and cleaning
up publicly available data, and then stores them in a local database or flat
files. This lets data analysts spend less time cleaning up and managing data,
and more time analyzing it.

This package lets you access the Data Retriever using R, so that the
Retriever's data handling can easily be integrated into R workflows.
}

