% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_with_widgets.R
\name{render_with_widgets}
\alias{render_with_widgets}
\title{Query a corpus for a topic and create a report summarizing and containing the results}
\usage{
render_with_widgets(input_file, output_file = sub("\\\\.Rmd$", ".html",
  input_file, ignore.case = TRUE), self_contained = TRUE,
  deps_path = file.path(dirname(output_file), "deps"))
}
\arguments{
\item{input_file}{Path to input file}
}
\description{
Query a corpus for a topic and create a report summarizing and containing the results
}
\examples{
render_with_widgets
}
\keyword{report}
