\name{retriever-package}
\alias{retriever-package}
\alias{retriever}
\docType{package}
\title{
\packageTitle{retriever}
}
\description{
\packageDescription{retriever}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{retriever}
\packageIndices{retriever}
}
\author{
\packageAuthor{retriever}
Maintainer: \packageMaintainer{retriever}
}
\references{}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{}
