\name{update}
\alias{update}
\title{Update the scripts the EcoData Retriever uses to download datasets}
\usage{
update()
}
\value{
returns the log of the Retriever's update
}
\description{
Update the scripts the EcoData Retriever uses to download
datasets
}
\examples{
ecoretriever::update()
}
\references{
http://ecodataretriever.org/cli.html
}

