\name{update_scripts}
\alias{update_scripts}
\title{Update the scripts the EcoData Retriever uses to download datasets}
\usage{
update_scripts()
}
\value{
returns the log of the Retriever's update
}
\description{
Update the scripts the EcoData Retriever uses to download
datasets
}
\examples{
update_scripts()
}
\references{
http://ecodataretriever.org/cli.html
}

