% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdataretriever.R
\name{datasets}
\alias{datasets}
\title{Name all available dataset scripts.}
\usage{
datasets()
}
\value{
returns a character vector with the available datasets for download
}
\description{
Additional information on the available datasets can be found at http://data-retriever.org/available-data.html
}
\examples{
\donttest{
rdataretriever::datasets()
}
}

