% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datasets}
\alias{datasets}
\title{Display a list all available dataset scripts.}
\usage{
datasets()
}
\value{
returns the log of the available datasets for download
}
\description{
Additional information on the available datasets can be found at http://ecodataretriever.org/available-data.html
}
\examples{
ecoretriever::datasets()
}

