% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecoretriever.R
\name{datasets}
\alias{datasets}
\title{Name all available dataset scripts.}
\usage{
datasets()
}
\value{
returns a character vector with the available datasets for download
}
\description{
Additional information on the available datasets can be found at http://ecodataretriever.org/available-data.html
}
\examples{
\donttest{
ecoretriever::datasets()
}
}

