\name{datasets}
\alias{datasets}
\title{Display a list all available dataset scripts}
\usage{
datasets()
}
\value{
returns the log of the available datasets for download
}
\description{
Display a list all available dataset scripts
}
\examples{
## Use an explicit namespace call when using datasets() to avoid conflicts
ecoretriever::datasets()
}

