% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_updates}
\alias{get_updates}
\title{Update the retrievers download script to the most recent official release.}
\usage{
get_updates()
}
\description{
This function will check if the version of retriever scripts in your local
directory ~/.retriever/scripts/ is up-to-date with the most recent offical
retriever release. Note it is possible that even more updated scripts exist
at the retriever repository https://github.com/weecology/retriever/tree/master/scripts
that have not yet been incorperated into an official release, and you should
consider checking that page if you have any concerns.
}
\examples{
ecoretriever::get_updates()
}

